Solution to the answer
You will be given an array of integers and a target value. Determine the number of pairs of array elements that have a difference equal to a target value.

For example, 
 
Given an array of [1, 2, 3, 4] and a target value of 1. The output should be (2, 1), (3, 2) and (4, 3).

We have three values meeting the condition: 
2-1=1,
3-2=1 and 4-3=1.